Move CategoryFilter to components dir
