+++ src/containers/BooksList.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Book from '../components/Book';
 import { removeBook, changeFilter } from '../actions';
-import CategoryFilter from './CategoryFilter';
+import CategoryFilter from '../components/CategoryFilter';
 
 const BooksList = ({ books, filter, dispatch }) => {
   const handleRemoveBook = ({ bookId }) => {
